+++ frontend/libs/altinn-icons/src/components/Accordion/index.ts
