+++ frontend/libs/altinn-icons/src/components/Checkbox/index.ts
