+++ frontend/libs/altinn-icons/src/components/Group/Group.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { IconProps } from '../../types/IconProps';
+import { SvgTemplate } from '../SvgTemplate';
 
 /**
  * @param {IconProps} props the icon props
@@ -7,17 +8,9 @@ import { IconProps } from '../../types/IconProps';
  * @example
  * import { Group } from '@altinn/icons';
  */
-export const Group = ({ title, ...rest }: IconProps): JSX.Element => {
+export const Group = (props: IconProps): JSX.Element => {
   return (
-    <svg
-      width='24'
-      height='24'
-      viewBox='0 0 24 24'
-      fill='none'
-      xmlns='http://www.w3.org/2000/svg'
-      {...rest}
-    >
-      {title && <title>{title}</title>}
+    <SvgTemplate {...props}>
       <rect
         x='4.42725'
         y='4.90283'
@@ -28,6 +21,6 @@ export const Group = ({ title, ...rest }: IconProps): JSX.Element => {
         strokeWidth='1.5'
         strokeDasharray='3 3'
       />
-    </svg>
+    </SvgTemplate>
   );
 };
