+++ frontend/libs/altinn-icons/src/components/Group/index.ts
