+++ frontend/libs/altinn-icons/src/components/Likert/index.ts
