+++ frontend/libs/altinn-icons/src/components/LongText/LongText.tsx
@@ -1,47 +1,49 @@
 import React from 'react';
 import { IconProps } from '../../types/IconProps';
+import { SvgTemplate } from '../SvgTemplate';
 
 /**
  * @param {IconProps} props the icon props
  * @returns {JSX.Element} the icon as a react component
  * @example
- * import { Paragraph } from '@altinn/icons';
+ * import { LongText } from '@altinn/icons';
  */
-export const Paragraph = ({ title, ...rest }: IconProps): JSX.Element => {
+export const LongText = (props: IconProps): JSX.Element => {
   return (
-    <svg
-      width='24'
-      height='24'
-      viewBox='0 0 24 24'
-      fill='none'
-      xmlns='http://www.w3.org/2000/svg'
-      {...rest}
-    >
-      {title && <title>{title}</title>}
+    <SvgTemplate {...props}>
+      <rect
+        x='3.95972'
+        y='4.76318'
+        width='18'
+        height='16'
+        rx='3'
+        stroke='currentColor'
+        strokeWidth='1.5'
+      />
       <path
-        d='M4.95947 18.6934H12.9595'
+        d='M6.95972 14.4292H18.9597'
         stroke='currentColor'
         strokeWidth='1.5'
         strokeLinecap='round'
       />
       <path
-        d='M4.95947 10.6436H12.9595'
+        d='M6.95972 11.2388H18.9597'
         stroke='currentColor'
         strokeWidth='1.5'
         strokeLinecap='round'
       />
       <path
-        d='M4.95947 14.6934H20.9595'
+        d='M6.95972 8.04834H18.9597'
         stroke='currentColor'
         strokeWidth='1.5'
         strokeLinecap='round'
       />
       <path
-        d='M4.95947 6.69336H20.9595'
+        d='M6.96069 17.4292H12.9607'
         stroke='currentColor'
         strokeWidth='1.5'
         strokeLinecap='round'
       />
-    </svg>
+    </SvgTemplate>
   );
 };
