+++ frontend/libs/altinn-icons/src/components/LongText/index.ts
