+++ frontend/libs/altinn-icons/src/components/NavBar/index.ts
