+++ frontend/libs/altinn-icons/src/components/Paragraph/Paragraph.tsx
@@ -1,56 +1,40 @@
 import React from 'react';
 import { IconProps } from '../../types/IconProps';
+import { SvgTemplate } from '../SvgTemplate';
 
 /**
  * @param {IconProps} props the icon props
  * @returns {JSX.Element} the icon as a react component
  * @example
- * import { LongText } from '@altinn/icons';
+ * import { Paragraph } from '@altinn/icons';
  */
-export const LongText = ({ title, ...rest }: IconProps): JSX.Element => {
+export const Paragraph = (props: IconProps): JSX.Element => {
   return (
-    <svg
-      width='24'
-      height='24'
-      viewBox='0 0 24 24'
-      fill='none'
-      xmlns='http://www.w3.org/2000/svg'
-      {...rest}
-    >
-      {title && <title>{title}</title>}
-      <rect
-        x='3.95972'
-        y='4.76318'
-        width='18'
-        height='16'
-        rx='3'
-        stroke='currentColor'
-        strokeWidth='1.5'
-      />
+    <SvgTemplate {...props}>
       <path
-        d='M6.95972 14.4292H18.9597'
+        d='M4.95947 18.6934H12.9595'
         stroke='currentColor'
         strokeWidth='1.5'
         strokeLinecap='round'
       />
       <path
-        d='M6.95972 11.2388H18.9597'
+        d='M4.95947 10.6436H12.9595'
         stroke='currentColor'
         strokeWidth='1.5'
         strokeLinecap='round'
       />
       <path
-        d='M6.95972 8.04834H18.9597'
+        d='M4.95947 14.6934H20.9595'
         stroke='currentColor'
         strokeWidth='1.5'
         strokeLinecap='round'
       />
       <path
-        d='M6.96069 17.4292H12.9607'
+        d='M4.95947 6.69336H20.9595'
         stroke='currentColor'
         strokeWidth='1.5'
         strokeLinecap='round'
       />
-    </svg>
+    </SvgTemplate>
   );
 };
