+++ frontend/libs/altinn-icons/src/components/Paragraph/index.ts
