+++ frontend/libs/altinn-icons/src/components/RadioButton/index.ts
