+++ frontend/libs/altinn-icons/src/components/Select/index.ts
