+++ frontend/libs/altinn-icons/src/components/ShortText/index.ts
