+++ frontend/libs/altinn-icons/src/components/SvgTemplate/index.ts
@@ -1,0 +1,1 @@
+export { SvgTemplate } from './SvgTemplate';
