+++ frontend/libs/altinn-icons/src/components/index.ts
